explosions added
explosion vfx added. I wanted to have a scene that handles explosions seperately, but it's really hard to get an area3D to update when you want it to (e.i. in the first frame it's instanced) so theres a seperated functionality that has been implemented into the missile and the grenade. couldn't find a punchy explosion sound for some reason.